Move PutRssUserConfigInsensitive to wicket_common
